+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## 0.9.0
 
+- Update Kamelet Catalog from 3.20.4 to 3.20.5
+
 ## 0.8.0
 
 - Minimal version of VS Code to run this extension is 1.76.0